requirements.txt
jupyter==1.0.0
pandas
numpy
selenium
urllib